cambios promiscuos

+++ src/main/java/DAW/lope/tienda/controllers/ControladorTienda.java
@@ -107,7 +107,7 @@ public class ControladorTienda {
 			@RequestParam String contrasenia,
 			@RequestParam String correo,
 			@RequestParam String fechanacimiento,
-			@RequestParam int numerotarjeta,
+			@RequestParam long numerotarjeta,
 			@RequestParam String titulartarjeta,
 			@RequestParam int codigoseguridad,
 			@RequestParam String direccionfacturacion) {
